clk: qcom: clk-cpu-osm: Limit the rate of CPU clock-rate changes
Without any form of rate limiting on CPU clock-rate changes, the CPU
clock can be changed as frequently as every 50 usecs, which is not
nearly enough time for the hardware to settle into the new rate. This
can result in the CPU performing sub-optimally when the clock-rate is
changed extremely frequently.

To mitigate this, rate-limit CPU clock-rate updates to once every 10 ms.
Additionally, remove clk_pwrcl_set_rate since its own delay mechanism is
now redundant.

Change-Id: Ia3ad18d388a0007552d2dac9a9980448640a585b
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>