clk: qcom: clk-cpu-osm: Don't sleep inside the global clk lock
The clk framework uses a global mutex for all clock operations. This
means that when one clock user has the lock held, no other clocks can be
changed during that time. This can result in slow clock-rate propagation
to unrelated clocks.

To remedy this, only sleep within a per-cluster mutex lock. This not
only allows different clusters to quickly update their clock rates
independent of one another, but also prevents stalls when non-CPU clocks
are updated, since the global clk lock will not be held for a long time.

Change-Id: I010c003557e4bed414b85fe51c6ec4899db43b1e
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>